INTERNAL: Focus manager renamed
