+++ src/gui/Components/Form/RadioButton.tsx
@@ -19,7 +19,7 @@ along with ORIGAM. If not, see <http://www.gnu.org/licenses/>.
 
 import React from "react";
 import S from "gui/Components/Form/RadioButton.module.scss";
-import {IFocusAble} from "model/entities/FocusManager";
+import {IFocusAble} from "model/entities/FormFocusManager";
 import { v4 as uuidv4 } from 'uuid';
