+++ src/gui/Components/ScreenElements/Editors/BoolEditor.tsx
@@ -21,7 +21,7 @@ import * as React from "react";
 import { observer } from "mobx-react";
 import S from "./BoolEditor.module.scss";
 import cx from "classnames";
-import { IFocusAble } from "../../../../model/entities/FocusManager";
+import { IFocusAble } from "../../../../model/entities/FormFocusManager";
 import CS from "gui/Components/ScreenElements/Editors/CommonStyle.module.css";
 
 @observer
