+++ src/gui/Components/ScreenElements/Editors/DropdownEditor.tsx
@@ -36,7 +36,7 @@ import { getMenuItemId } from "../../../../model/selectors/getMenuItemId";
 import { Dropdowner } from "gui/Components/Dropdowner/Dropdowner";
 import { getEntity } from "../../../../model/selectors/DataView/getEntity";
 import { getSessionId } from "model/selectors/getSessionId";
-import { IFocusAble } from "../../../../model/entities/FocusManager";
+import { IFocusAble } from "../../../../model/entities/FormFocusManager";
 import { rowHeight } from "gui/Components/ScreenElements/Table/TableRendering/cells/cellsCommon";
 
 export interface IDropdownEditorProps {
