+++ src/gui/Components/ScreenElements/Editors/TextEditor.tsx
@@ -21,7 +21,7 @@ import { action } from "mobx";
 import { observer } from "mobx-react";
 import * as React from "react";
 import S from "./TextEditor.module.scss";
-import { IFocusAble } from "../../../../model/entities/FocusManager";
+import { IFocusAble } from "../../../../model/entities/FormFocusManager";
 
 import { EditorState, convertToRaw, ContentState } from "draft-js";
 import { Editor } from "react-draft-wysiwyg";
