+++ src/model/selectors/DataView/getFormFocusManager.ts
@@ -17,9 +17,9 @@ You should have received a copy of the GNU General Public License
 along with ORIGAM. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { FocusManager } from "model/entities/FocusManager";
+import { FormFocusManager } from "model/entities/FormFocusManager";
 import { getDataView } from "model/selectors/DataView/getDataView";
 
-export function getFocusManager(ctx: any): FocusManager {
-  return getDataView(ctx).focusManager;
+export function getFormFocusManager(ctx: any): FormFocusManager {
+  return getDataView(ctx).formFocusManager;
 }
